git-svn-id: http://lpm.org.ru/svn/kumir2@424 382dccb8-6cdf-11e0-9640-eb60159f4418

+++ src/plugins/editor/editorplane.cpp
@@ -298,7 +298,7 @@ void EditorPlane::mouseMoveEvent(QMouseEvent *e)
         QApplication::restoreOverrideCursor();
         int dX = e->pos().x() - pnt_textPress.x();
         int dY = e->pos().y() - pnt_textPress.y();
-        qreal distance = sqrt(dX*dX+dY*dY);
+        qreal distance = sqrt(double(dX*dX+dY*dY));
         if (pnt_textPress.x()==-1000 && !b_selectionInProgress)
             distance = 0;
         qreal therehold = QApplication::startDragDistance();
